% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{SubtypeSpecificGene}
\alias{SubtypeSpecificGene}
\title{Tool for Identification of subtype-specific genes}
\usage{
SubtypeSpecificGene(omics = NULL, cluster = NULL)
}
\arguments{
\item{omics, cluster}{}
}
\description{
Given a gene from a list of genes of interest, it will be specifically distributed to either of the identified subgroups based on its mean values (e.g., CNA changes, MET changes, and expression levels). Then, a gene was considered as a subtype-specific one if P-value <= 0.05 (one-way ANOVA test). For further information on requirements as well as how to implement this tool, please visit my Github repository: https://github.com/huynguyen250896/GeneCluster
}
\examples{
SubtypeSpecificGene(omics = df, cluster = groups)

}
